Renderer/pyproject.toml
[tool.poetry]
name = "renderer"
version = "0.1.0"
description = "Data visualization for system metrics"
authors = ["CPKCShanKonduru <CPKCShanKonduru@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
matplotlib = "^3.8.0"
psycopg2-binary = "^2.9.9"
tk = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
